0513 코드리뷰

See merge request ai_track/class_06/ai_project/team03/frontend!3

+++ main-app/src/pages/AdminPage/UserManagePage.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useGetUsers, useDeleteUser } from '../../API/AdminUserApi';
 import styled from 'styled-components';
+import Loading from '../../components/common/Loading';
 
 function UserManage() {
   const [users, setUsers] = useState([]);
@@ -14,27 +15,20 @@ function UserManage() {
   }, [fetchedUsers, setUsers]);
 
   const handleDelete = async (id) => {
-    try {
-      const response = await deleteUser(id);
-      setUsers((prevUsers) =>
-        prevUsers.map((user) => {
-          if (user.id === id) {
-            return {
-              ...user,
-              deletedAt: new Date(),
-            };
-          }
-          return user;
-        })
-      );
-      alert(response.data.message);
-    } catch (err) {
-      console.log(err.message);
+    if (window.confirm('삭제하시겠습니까?')) {
+      try {
+        await deleteUser(id); // 서버에서 데이터를 소프트 딜리트
+        setUsers((prevUsers) => prevUsers.filter((user) => user.id !== id)); // 클라이언트에서 실제 데이터를 제거
+        alert('삭제되었습니다.');
+      } catch (err) {
+        console.log(err.message);
+        alert('삭제에 실패했습니다.');
+      }
     }
   };
 
   if (isLoading) {
-    return <div>Loading...</div>;
+    return <Loading />;
   }
 
   if (error) {
@@ -45,9 +39,8 @@ function UserManage() {
     <Table>
       <thead>
         <tr>
-          <TableHeader>Created At</TableHeader>
+          <TableHeader>가입일</TableHeader>
           <TableHeader>이메일</TableHeader>
-          <TableHeader>이름</TableHeader>
           <TableHeader>닉네임</TableHeader>
           <TableHeader>유형</TableHeader>
           <TableHeader>관리</TableHeader>
@@ -57,11 +50,12 @@ function UserManage() {
         {users &&
           users.map((user) => (
             <tr key={user.id}>
-              <TableData>{user.createdAt}</TableData>
+              <TableData style={{ width: '13%' }}>
+                {new Date(user.created_at).toISOString().substring(0, 10)}
+              </TableData>
               <TableData>{user.email}</TableData>
-              <TableData>{user.name}</TableData>
               <TableData>{user.nickname}</TableData>
-              <TableData>{user.type}</TableData>
+              <TableData>{user.is_admin === 1 ? '어드민' : '회원'}</TableData>
               <TableData>
                 <DeleteButton
                   key={user.id}
